# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KL Project API — Multi-branch KTV management platform
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
  db?: string;
  timestamp?: string;
}

export interface ApiError {
  error: string;
  message?: string;
}

export interface ApiSuccess {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserProfile {
  id: string;
  name: string;
  email?: string;
  role: string;
  branchId?: string | null;
  userType: string;
}

export interface LoginResponse {
  accessToken: string;
  refreshToken: string;
  user: UserProfile;
}

export interface RefreshRequest {
  refreshToken: string;
}

export interface RefreshResponse {
  accessToken: string;
}

export type BranchTaxConfig = { [key: string]: unknown };

export type BranchSettings = { [key: string]: unknown };

export interface Branch {
  id: string;
  orgId: string;
  name: string;
  internalCode: string;
  address?: string | null;
  city?: string | null;
  country?: string | null;
  phone?: string | null;
  email?: string | null;
  timezone: string;
  currency: string;
  isActive: boolean;
  taxConfig?: BranchTaxConfig;
  settings?: BranchSettings;
  createdAt?: string;
}

export interface BranchResponse {
  data: Branch;
}

export interface BranchListResponse {
  data: Branch[];
}

export interface CreateBranchRequest {
  name: string;
  internalCode: string;
  orgId: string;
  address?: string;
  city?: string;
  country?: string;
  phone?: string;
  email?: string;
  timezone?: string;
  currency?: string;
}

export type UpdateBranchRequestTaxConfig = { [key: string]: unknown };

export type UpdateBranchRequestSettings = { [key: string]: unknown };

export interface UpdateBranchRequest {
  name?: string;
  address?: string;
  city?: string;
  phone?: string;
  email?: string;
  isActive?: boolean;
  taxConfig?: UpdateBranchRequestTaxConfig;
  settings?: UpdateBranchRequestSettings;
}

export type BranchDashboardToday = {
  revenue: number;
  reservationCount: number;
  checkedInCount: number;
  occupancyPct: number;
};

export type BranchDashboardRooms = {
  total: number;
  available: number;
  occupied: number;
  cleaning: number;
  maintenance: number;
};

export type BranchDashboardArrivalsItem = { [key: string]: unknown };

export type BranchDashboardDeparturesItem = { [key: string]: unknown };

export interface BranchDashboard {
  branchId: string;
  branchName: string;
  localTime: string;
  today: BranchDashboardToday;
  rooms: BranchDashboardRooms;
  arrivals?: BranchDashboardArrivalsItem[];
  departures?: BranchDashboardDeparturesItem[];
}

export type RoomRoomType = (typeof RoomRoomType)[keyof typeof RoomRoomType];

export const RoomRoomType = {
  private_room: "private_room",
  vip_room: "vip_room",
  vvip_room: "vvip_room",
  table: "table",
  open_area: "open_area",
} as const;

export type RoomStatus = (typeof RoomStatus)[keyof typeof RoomStatus];

export const RoomStatus = {
  available: "available",
  occupied: "occupied",
  cleaning: "cleaning",
  maintenance: "maintenance",
  blocked: "blocked",
} as const;

export interface Room {
  id: string;
  branchId: string;
  name: string;
  roomType: RoomRoomType;
  capacityMin: number;
  capacityMax: number;
  hourlyRate?: number | null;
  minHours?: number;
  status: RoomStatus;
  sortOrder?: number;
  isActive: boolean;
  description?: string | null;
  floorLevel?: string | null;
  createdAt?: string;
}

export type RoomWithReservation = Room & {
  reservationNo?: string | null;
  guestName?: string | null;
  guestCount?: number | null;
  checkInTime?: string | null;
  expectedCheckOut?: string | null;
};

export interface RoomResponse {
  data: Room;
}

export interface RoomListResponse {
  data: RoomWithReservation[];
}

export interface RoomBoardResponse {
  branchId: string;
  branchName: string;
  rooms: RoomWithReservation[];
  updatedAt: string;
}

export interface CreateRoomRequest {
  branchId: string;
  name: string;
  roomType?: string;
  capacityMin?: number;
  capacityMax: number;
  hourlyRate?: number;
  minHours?: number;
  description?: string;
  floorLevel?: string;
  sortOrder?: number;
}

export interface UpdateRoomRequest {
  name?: string;
  roomType?: string;
  capacityMin?: number;
  capacityMax?: number;
  hourlyRate?: number;
  minHours?: number;
  description?: string;
  floorLevel?: string;
  sortOrder?: number;
  isActive?: boolean;
}

export type UpdateRoomStatusRequestStatus =
  (typeof UpdateRoomStatusRequestStatus)[keyof typeof UpdateRoomStatusRequestStatus];

export const UpdateRoomStatusRequestStatus = {
  available: "available",
  occupied: "occupied",
  cleaning: "cleaning",
  maintenance: "maintenance",
  blocked: "blocked",
} as const;

export interface UpdateRoomStatusRequest {
  status: UpdateRoomStatusRequestStatus;
}

export interface MultilingualName {
  en: string;
  zh?: string | null;
  ms?: string | null;
  ko?: string | null;
  ja?: string | null;
  th?: string | null;
}

export interface ProductGroup {
  id: string;
  orgId: string;
  name: MultilingualName;
  sortOrder?: number;
  isActive: boolean;
  createdAt?: string;
}

export interface ProductGroupResponse {
  data: ProductGroup;
}

export interface ProductGroupListResponse {
  data: ProductGroup[];
}

export interface CreateProductGroupRequest {
  orgId: string;
  name: MultilingualName;
  sortOrder?: number;
}

export interface ProductType {
  id: string;
  groupId: string;
  name: MultilingualName;
  sortOrder?: number;
  isActive: boolean;
  createdAt?: string;
}

export interface ProductTypeResponse {
  data: ProductType;
}

export interface ProductTypeListResponse {
  data: ProductType[];
}

export interface CreateProductTypeRequest {
  groupId: string;
  name: MultilingualName;
  sortOrder?: number;
}

export type ProductDescription = { [key: string]: unknown } | null;

export interface Product {
  id: string;
  typeId: string;
  branchId?: string | null;
  sku?: string | null;
  name: MultilingualName;
  description?: ProductDescription;
  unitPrice: number;
  unit: string;
  taxApplicable?: boolean;
  sortOrder?: number;
  isActive: boolean;
  createdAt?: string;
}

export interface ProductResponse {
  data: Product;
}

export interface ProductListResponse {
  data: Product[];
}

export interface CreateProductRequest {
  typeId: string;
  branchId?: string;
  sku?: string;
  name: MultilingualName;
  unitPrice: number;
  unit?: string;
  taxApplicable?: boolean;
  sortOrder?: number;
}

export interface UpdateProductRequest {
  name?: MultilingualName;
  sku?: string;
  unitPrice?: number;
  unit?: string;
  taxApplicable?: boolean;
  sortOrder?: number;
  isActive?: boolean;
}

export type ListRoomsParams = {
  branch_id?: string;
};

export type ListAvailableRoomsParams = {
  branch_id: string;
  date?: string;
  start_time?: string;
  duration_hours?: number;
};

export type ListProductTypesParams = {
  group_id?: string;
};

export type ListProductsParams = {
  branch_id?: string;
  type_id?: string;
};