# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KL Project API — Multi-branch KTV management platform
 * OpenAPI spec version: 0.2.0
 */

export interface ApiError {
  error: string;
  message?: string;
}